Rename `vscode/src/viz.ts` to the clearer `vscode/src/webview-panel.ts`
